+++ noivaciaAPP/EspacoBase.cs
@@ -1,7 +1,7 @@
   public class EspacoBase{
-     public string Nome {get; set;}
+    public string Nome {get; set;}
     public int CapacidadeMax {get; set;}
-    public double Valor {get; set;}
+    public int Valor {get; set;}
     public EspacoTipoEnum Tipo {get; set;}
 
 }
